Introduce the repo manifest xml files
The Android R&D, which we are building should use the
common-android14-6.1 version of the Android kernel. That one is built
with the help of Bazel and repo. Repo is based on usage of the xml
manifest files.

This patch is introducing several xml repo manifest files, which
contain links to all required projects that are needed to build the
Android kernel.

That includes the xen-virtual-device repository, which is located in
xen-troops and is introducing the Xen-specific Bazel configuration.

The logic behind the introduced files is the following one:

- default.xml - entry point to include other files and locate the common
declarations
- common-android14-6.1.xml - the original Android kernel repo manifest,
which was fetched from its original location
- common-android14-6.1-override.xml - anything we add on top of the
original manifest

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>